Moving addLiquidity, addLiquidityNested, removeLiquidity and swap examples to v2 folder; Changing README.md file and adding structure;
